+++ lib/hooks/use-toast.ts
@@ -73,6 +73,16 @@ const addToRemoveQueue = (toastId: string) => {
 export const reducer = (state: State, action: Action): State => {
   switch (action.type) {
     case 'ADD_TOAST':
+      const exists = state.toasts.some((t) => t.id === action.toast.id);
+      if (exists) {
+        return {
+          ...state,
+          toasts: state.toasts.map((t) =>
+            t.id === action.toast.id ? { ...t, ...action.toast } : t,
+          ),
+        };
+      }
+
       return {
         ...state,
         toasts: [action.toast, ...state.toasts].slice(0, TOAST_LIMIT),
@@ -134,10 +144,10 @@ function dispatch(action: Action) {
   });
 }
 
-type Toast = Omit<ToasterToast, 'id'>;
+type Toast = Omit<ToasterToast, 'id'> & { id?: string };
 
 function toast({ ...props }: Toast) {
-  const id = genId();
+  const id = props.id || genId();
 
   const update = (props: ToasterToast) =>
     dispatch({